+++ lib/puppet_x/racadm/racadm.rb
@@ -52,4 +52,40 @@ class Racadm
   def self.s_to_bool(value)
     'Enabled'.eql? value
   end
+
+  def self.s_to_role(value)
+    case value
+    when '1'
+      'callback'
+    when '2'
+      'user'
+    when '3'
+      'operator'
+    when '4'
+      'administrator'
+    when '5'
+      'oem_proprietary'
+    when '15'
+      'no_access'
+    end
+  end
+
+  def self.role_to_s(value)
+    case value
+    when 'none'
+      '0'
+    when 'callback'
+      '1'
+    when 'user'
+      '2'
+    when 'operator'
+      '3'
+    when 'administrator'
+      '4'
+    when 'oem_proprietary'
+      '5'
+    when 'no_access'
+      '15'
+    end
+  end
 end